+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Github release action
+- Pre-Commit check
+
+### Fixed
+- Clippy suggestions
+- Formatting
 
 ## [v0.1.0] - 2026-01-25